move src/pageComponents/mint to app/mint/_components
