+++ template/web/app/mint/_components/ContractDemo.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useMemo, useState } from 'react';
 import { encodeFunctionData, formatEther } from 'viem';
 import { useAccount, useEstimateGas } from 'wagmi';
