+++ template/web/app/mint/_components/Guide.tsx
