+++ template/web/app/mint/_components/SwitchNetwork.tsx
