+++ template/web/app/mint/_components/steps/StartMintStep.tsx
